Add conditional stack rendering
