+++ src/components/AuthContext.js
@@ -8,13 +8,28 @@ export function useAuth() {
 }
 
 export function AuthProvider(props) {
+  /* Firebase user
+  This user is NOT from the database, it is the user from Firebase Auth
+  */
   const [currentUser, setCurrentUser] = useState();
+
+  useEffect(() => {
+    const unsubscribe = auth.onAuthStateChanged(onAuthStateChanged);
+    return () => unsubscribe();
+  }, []);
+
+  function onAuthStateChanged(user) {
+    setCurrentUser(user);
+  }
+
+  const isAuthenticated = !!currentUser;
 
   function signup(email, password) {
     return auth.createUserWithEmailAndPassword(email, password);
   }
 
   function login(email, password) {
+    console.log("trying to sign in");
     return auth.signInWithEmailAndPassword(email, password);
   }
 
@@ -26,15 +41,9 @@ export function AuthProvider(props) {
     return auth.sendPasswordResetEmail(email);
   }
 
-  useEffect(() => {
-    const unsubscribe = auth.onAuthStateChanged((user) => {
-      setCurrentUser(user);
-    });
-    return unsubscribe;
-  }, []);
-
   const value = {
     currentUser,
+    isAuthenticated,
     signup,
     login,
     logout,
